Update comparison.md
I added ribose to the popular projects table, if only to provoke some rethinking about the utility of the underlying technology, which derives more or less directly from work done by Thompson et al in the early unix days. The syntax is cranky (a problem that some of the frameworks above work to ameliorate) and they present limited support for fine-grained interaction with high-level code. Ribose is a ship-in-a-bottle demonstration of how [ginr](https://github.com/ntozubod/ginr), an industrial strength compiler for multidimensional regular expressions, can be harnessed to compile multitape transducers mapping input symbols (bytes) to non-branching vectors of simple effector methods expressed for a task-specfic object model. 

Ginr is the real gem in the ribose framework, but I'm not trying to sell anything here other than the general idea that there is a better way to harness (real) regular expressions for stream processing tasks ...